fix: silence GCC warnings in the skinned-mesh physics core
MSVC reports none of these, so the /W4 /WX Windows build stays green
either way, but a GCC build of the same code reports all of them:

- Vertex(): ZeroMemory over a struct holding a btVector3, which has a
  non-trivial constructor (-Wclass-memaccess). The member init list
  covers the whole struct.
- ColliderTree: the members filled in while the tree is built were
  indeterminate until then, and trees are copied before that
  (-Wmaybe-uninitialized).
- BoneScaleConstraint: mem-init list ordered like the declarations
  (-Wreorder).
- dispatch(): result.normOnB defined up front, for the negation in the
  swapping addResult (-Wmaybe-uninitialized).
- XMLReader: the Inspected values the two switches don't act on
  (-Wswitch).
- size_t loop counters where the bound is a std::vector size
  (-Wsign-compare).

No behavior change.

+++ src/hdtSkinnedMesh/hdtSkinnedMeshBody.cpp
@@ -219,7 +219,7 @@ namespace hdt
 
 		UINT numUsed = 0;
 		std::vector<UINT> map(m_vertices.size());
-		for (int i = 0; i < m_vertices.size(); ++i) {
+		for (size_t i = 0; i < m_vertices.size(); ++i) {
 			if (flags[i]) {
 				m_vertices[numUsed] = m_vertices[i];
 				m_vpos[numUsed] = m_vpos[i];
